prediction/requirements.txt
tensorflow>=2.12.0
numpy>=1.21.0
thrift>=0.16.0
pure-sasl>=0.6.2
thrift-sasl>=0.4.3
pyhive>=0.6.5
cassandra-driver>=3.25.0